cleaned up main.c
