+++ Makefile
@@ -1,4 +1,4 @@
-PIC_CC=/home/shared/Dropbox/Wifly_Light/Pic_program_pb/cc5xfree/CC5x.EXE
+PIC_CC=/home/gpb/cc5xfree/CC5x.EXE
 INC_DIR=include_files
 
 all:
@@ -8,7 +8,7 @@ pic:
 	wine ${PIC_CC} main.c -Iinclude_files
 
 x86:
-	gcc main.c commandstorage.c eeprom.c ledstrip.c RingBuf.c spi.c usart.c x86_wrapper.c -DX86 -I${INC_DIR}
+	gcc main.c commandstorage.c eeprom.c error.c ledstrip.c RingBuf.c spi.c usart.c x86_wrapper.c -DX86 -I${INC_DIR}
 
 clean:
 	rm *.hex *.occ
