+++ commandstorage.c
@@ -50,7 +50,7 @@ void commandstorage_get_commands()
 	{
 		// *** if a RingBufError occure, I have to throw away the current command,
 		// *** because the last byte was not saved. Commandstring is inconsistent
-		ClearCmdBuf;
+		ClearCmdBuf();
 	}
 
 	if(RingBufIsNotEmpty)
@@ -182,3 +182,17 @@ void commandstorage_execute_commands()
 		}
 	}
 }
+
+void commandstorage_init()
+{
+	/** EEPROM contains FF in every cell after inital start,
+	*** so I have to delete the pointer address
+	*** otherwise the PIC thinks he has the EEPROM full with commands
+	**/
+	if (EEPROM_RD(CmdPointerAddr) == 0xff)
+		EEPROM_WR(CmdPointerAddr, 0);
+
+	// set loop pointer address to start
+	EEPROM_WR(CmdLoopPointerAddr, 0);
+}
+
