+++ error.c
@@ -1,0 +1,28 @@
+
+#include "commandstorage.h"
+#include "error.h"
+#include "RingBuf.h"
+#include "usart.h"
+
+void throw_errors()
+{
+	if(RingBufHasError) 
+	{
+		// *** if a RingBufError occure, I have to throw away the current command,
+		// *** because the last byte was not saved. Commandstring is inconsistent
+		ClearCmdBuf();
+		USARTsend_str(" ERROR: Receivebuffer full");
+		// *** Re-init the Ringbuffer to get a consistent commandstring and reset error
+		RingBufInit();
+	}
+	if(gERROR.crc_failure)
+	{
+		USARTsend_str(" ERROR: CRC-Check failed");
+		gERROR.crc_failure = 0;
+	}
+	if(gERROR.eeprom_failure)
+	{
+		USARTsend_str(" ERROR: EEPROM is full");
+		gERROR.eeprom_failure = 0;
+	}
+}
